backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
pydantic>=2.6
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
passlib[argon2]>=1.7
argon2-cffi>=23.1.0
python-multipart>=0.0.9
pytest>=8.0
httpx>=0.27
pydantic[email]
